% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GETBIN}
\alias{GETBIN}
\title{Generate binary bivariate data.}
\usage{
GETBIN(PVEC)
}
\arguments{
\item{PVEC}{Contains TRUE pi00,pi10, pi01, pi11 for each dose}
}
\value{
Binary bivariate random variate (YE,YT).
}
\description{
Generates binary bivariate data from a 2x2 matrix of probabilities.
}
