% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReturnOpt}
\alias{ReturnOpt}
\title{Obtains optimal dose level.}
\usage{
ReturnOpt(Desire, DoseTried)
}
\arguments{
\item{Desire}{Vector of optimality function values for each dose level.}

\item{DoseTried}{Binary vector containing 1 for doses that have been tried during the trial and 0 otherwise.}
}
\value{
Optimal dose level to assign the next patient cohort.
}
\description{
Gives the optimal dose level to assign the next patient cohort to based on the vector of optimality function values for each dose level (desirability or utility based) and the doses tried throughout the trial.
}
